Conversion from partial waves to operator basis
New subroutine to convert the potential parameters in partial wave basis
to operators basis as defined in Appendix A of PhysRevC.88.064002 (Table X)

To check that the conversion is correct, a subroutine to calculate any
partial wave (given the corresponding quantum numbers) from the operator basis
was written. The original partial waves from the input file are recovered.

Signed-off-by: "Rodrigo Navarro Perez" <"rnavarroperez@sdsu.edu">